% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_additive_mod_wand.R
\name{add_wand_gen_additive_mod}
\alias{add_wand_gen_additive_mod}
\title{Wrapper to add \code{wand} as an engine to parsnip's \code{gen_additive_model} model spec}
\usage{
add_wand_gen_additive_mod()
}
\description{
Wrapper to add \code{wand} as an engine to parsnip's \code{gen_additive_model} model spec
}
