% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_wand_dataset.R
\name{build_wand_dataset}
\alias{build_wand_dataset}
\title{Builds \code{wand} compatible \code{torch::tensor_dataset}s}
\usage{
build_wand_dataset(
  linear_predictors,
  smooth_predictors,
  outcome,
  requires_grad = T
)
}
\arguments{
\item{linear_predictors}{The linear predictors as an R \strong{dataframe} or \strong{matrix}.}

\item{smooth_predictors}{A named list of R where each entry is a \strong{dataframe} or \strong{matrix}
representing a smooth predictor.}

\item{requires_grad}{A boolean indicating whether or not the prediction tensors need gradients.
This can be turned off for datasets that are only used for evaluation, e.g. validation data or
prediction data.}

\item{y}{The target as an R \strong{vector}.}
}
\description{
Builds \code{wand} compatible \code{torch::tensor_dataset}s
}
