% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{build_wand_graph}
\alias{build_wand_graph}
\title{Builds a \code{ggraph::tbl_graph} showing data flow through a wand model}
\usage{
build_wand_graph(wand_fit)
}
\arguments{
\item{wand_fit}{A fitted \code{wand} model.}
}
\value{
A \code{ggraph::tbl_graph} graph of the model.
}
\description{
Builds a \code{ggraph::tbl_graph} showing data flow through a wand model
}
