% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_inspection.R
\name{coef.wand}
\alias{coef.wand}
\title{Extract model coefficients}
\usage{
\method{coef}{wand}(wand_fit, ...)
}
\arguments{
\item{wand_fit}{A fitted \code{wand} model.}
}
\value{
A matrix of the final linear layer's weights and biases. Note that classification models
use a softmax activation, which returns one column per class.
}
\description{
Extract model coefficients
}
