% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dehydrate_model}
\alias{dehydrate_model}
\title{Store a \code{torch} module as a raw R object.}
\usage{
dehydrate_model(model)
}
\arguments{
\item{model}{A \code{torch} module.}
}
\value{
A raw object.
}
\description{
This function is used to store \code{torch} modules, usually after training, as raw R objects. This
function comes directly from the \code{brulee} package.
}
