% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dots_to_formula}
\alias{dots_to_formula}
\title{Combine ... into a RHS only formula}
\usage{
dots_to_formula(...)
}
\arguments{
\item{...}{An unquoted set of terms to be combined into a formula.}
}
\value{
A RHS-only formula generated by just adding the ... terms.
}
\description{
Combine ... into a RHS only formula
}
