% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_smooths}
\alias{extract_smooths}
\title{Takes a formula, and replaces any smooth specs with the corresponding formula. Returns the
updated formula and the smooth specs extracted from the formula}
\usage{
extract_smooths(formula)
}
\arguments{
\item{formula}{A formula that may contain \code{wand} \code{s_} functions.}
}
\value{
A named list with a \code{formula} and \code{smooth_specs} entries. The \code{formula} matches the input
formula, but with all \code{s_} functions replaced with just their contents. The \code{smooth_specs}
entry contains a list of smooth specifications.
}
\description{
Takes a formula, and replaces any smooth specs with the corresponding formula. Returns the
updated formula and the smooth specs extracted from the formula
}
