% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_var_spacing}
\alias{get_var_spacing}
\title{Get average 1d distance between points}
\usage{
get_var_spacing(var)
}
\arguments{
\item{var}{A vector.}
}
\value{
If \code{var} is numeric, then this returns the average distance between \code{var} points. If
\code{var} is a character or factor, then \code{get_var_spacing} just returns the unique levels.
}
\description{
Get average 1d distance between points
}
