% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hydrate_model}
\alias{hydrate_model}
\title{Load models from raw objects.}
\usage{
hydrate_model(model)
}
\arguments{
\item{model}{A raw object representing a \code{torch} module.}
}
\value{
A \code{torch} module.
}
\description{
This function is used to restore dehydrated \code{torch} modules from raw R objects. This function
comes directly from the \code{brulee} package.
}
