% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{incant}
\alias{incant}
\title{Return file info}
\usage{
incant(path)
}
\arguments{
\item{path}{character vector of files to use magic on}
}
\value{
a \code{tibble} / \code{data.frame} of file magic attributes
}
\description{
Return file info
}
\examples{
library(magrittr)
library(dplyr)

system.file("img", package="filemagic") \%>\%
  list.files(full.names=TRUE) \%>\%
  incant() \%>\%
  glimpse()
}

