% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_additive_mod_wand.R
\name{nn_additive_mod}
\alias{nn_additive_mod}
\title{Generalized additive model via neural networks}
\usage{
nn_additive_mod(
  mode = "unknown",
  engine = "wand",
  batch_size = NULL,
  epochs = NULL,
  learn_rate = NULL,
  stop_iter = NULL
)
}
\arguments{
\item{mode}{A single character string for the prediction outcome mode. Possible values for this
model are "unknown", "regression", or "classification".}

\item{engine}{A single character string specifying what computational engine to use for fitting.}

\item{batch_size}{An integer giving the number of training samples included in each minibatch.}

\item{epochs}{An integer giving the number of training epochs.}

\item{learn_rate}{The initial learning rate used by the optimizer.}

\item{stop_iter}{A non-negative integer giving the number of epochs with no improvement in loss
before training is stopped. If \code{validation_prop > 0} then validation loss is used, otherwise
loss is evaluated using the whole training set. Set this parameter to \code{Inf} to prevent any
early stopping.}
}
\description{
Generalized additive model via neural networks
}
