% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_specifiers.R
\name{s_mlp}
\alias{s_mlp}
\title{Specify a smooth function using a multilayer perceptron}
\usage{
s_mlp(..., hidden_units = c(32, 32, 32))
}
\arguments{
\item{...}{The unquoted variables to be smoothed.}

\item{hidden_units}{An integer vector specifying the number of hidden units in each layer of the
mlp. Note that \code{length(hidden_units)} is the number of layers in the mlp; the final entry in
\code{hidden_units} represents the output of this smoother.}
}
\value{
A named list containing the features to be smoothed, the torch module doing the smoothing,
the parameters needed to initialize that module, and the number of features returned by the mlp.
}
\description{
This function specifies a smooth function using a multilayer perceptron (mlp). Note that this
function does not do the actual smoothing, it merely specifies the features to be smoothed,
the \code{torch::nn_module} module to be used to do the smoothing, in this case an MLP, and the
parameters needed to initialize that module.
}
