% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_y}
\alias{scale_y}
\title{Scales an outcome given a list containing mean and sd.}
\usage{
scale_y(y, stats)
}
\arguments{
\item{y}{The numeric outcome of interest.}

\item{stats}{A named list containing \code{mean} and \code{sd} entries.}
}
\value{
The scaled \code{y} value
}
\description{
This function comes from brulee. It is used to scale the outcome before training and prediction.
}
