% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.r
\docType{data}
\name{simplemagic_mime_db}
\alias{simplemagic_mime_db}
\title{File extension-to-MIME mapping data frame}
\description{
built from the files found in the References
}
\note{
Last updated 2019-07-27.
}
\references{
- <https://tools.ietf.org/html/rfc6838>
    - <https://www.iana.org/assignments/media-types/media-types.xhtml>
    - <https://s-randomfiles.s3.amazonaws.com/mime/allMimeTypes.json>
    - <https://cdn.jsdelivr.net/gh/jshttp/mime-db@master/db.json>
}
\keyword{datasets}
