% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{spacing}
\alias{spacing}
\title{Get mean 1d distance between points}
\usage{
spacing(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The average distance between \code{x} points.
}
\description{
Get mean 1d distance between points
}
