% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarise_predictor}
\alias{summarise_predictor}
\title{Get summary statistics for a}
\usage{
summarise_predictor(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A named list of summary statistics.
}
\description{
Get summary statistics for a
}
