% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wand-package.R
\docType{package}
\name{wand-package}
\alias{wand-package}
\alias{_PACKAGE}
\title{wand: Fit Wide and Deep Neural Networks}
\description{
Provides functions for fitting and examining wide and deep neural networks in a {tidymodels} friednly framework. Powered by {torch}.
}
\author{
\strong{Maintainer}: Ben Ewing \email{ben.j.ewing@protonmail.com}

}
\keyword{internal}
