% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wand-package.R
\docType{package}
\name{wand}
\alias{wand}
\alias{wand-package}
\title{Retrieve 'Magic' Attributes from Files and Directories}
\description{
Retrieve 'Magic' Attributes from Files and Directories
}
\author{
Bob Rudis (@hrbrmstr)
}

