% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_modules.R
\name{wand_classification_module}
\alias{wand_classification_module}
\title{A linear regression \code{torch::nn_module} module for classification}
\usage{
wand_classification_module(n_features, n_classes)
}
\arguments{
\item{n_features}{An integer giving the number of input features. Used to initialize the module.}

\item{n_classes}{An integer giving the number of output classes Used to initialize the module.}
}
\value{
A \code{torch::nn_module} module which returns a returns \code{n_classes} values using a
\code{torch::nn_softmax} activation.
}
\description{
This module takes in \code{n_features} and returns \code{n_classes} values using a \code{torch::nn_softmax}
activation.
}
