% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_inspection.R
\name{wand_plot_loss}
\alias{wand_plot_loss}
\title{Plot model loss by epoch}
\usage{
wand_plot_loss(
  wand_fit,
  show_best_epoch = T,
  show_early_stopping = T,
  show_labels = T
)
}
\arguments{
\item{wand_fit}{A fitted \code{wand} model.}

\item{show_best_epoch}{Logical indicating whether or not a vline indicating the best training
epoch should be included. If \code{validation_prop > 0} then this line will identify the best
validation epoch, otherwise it will identify the best training epoch.}

\item{show_early_stopping}{Logical indicating whether or not a vline indicating the early
stopping point should be included.}

\item{show_labels}{Logical indicating whether or not annotations labeling the best epoch and
early stopping lines (if they are enabled) should be included.}
}
\value{
A ggplot.
}
\description{
Plot model loss by epoch
}
