% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_inspection.R
\name{wand_plot_smooths}
\alias{wand_plot_smooths}
\title{Plot smooth functions from a wand model}
\usage{
wand_plot_smooths(wand_fit, df, seq_length = 250)
}
\arguments{
\item{wand_fit}{A fitted \code{wand} model.}

\item{df}{A dataframe used to determine feature ranges and spacing. It is recommended to use the
training dataframe.}

\item{seq_length}{The maximum number of points used when plotting each smooth.}
}
\value{
A list where each entry is a ggplot of a smooth function.
}
\description{
Plot smooth functions from a wand model
}
