% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_modules.R
\name{wand_regression_module}
\alias{wand_regression_module}
\title{TODO should s_ functions be an S3 class?
TODO what about an s_ constructor function?
A linear regression \code{torch::nn_module} module}
\usage{
wand_regression_module(n_features)
}
\arguments{
\item{n_features}{An integer giving the number of input features. Used to initialize the module.}
}
\value{
A \code{torch::nn_module} module which returns a single feature with no activation.
}
\description{
This module takes in \code{n_features} and returns a single feature, with no activation.
}
