% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_wand_dataset}
\alias{build_wand_dataset}
\title{Builds \code{wand} compatible \code{torch::tensor_dataset}s}
\usage{
build_wand_dataset(x, y, smooth_specs, requires_grad = T)
}
\arguments{
\item{x}{The predictors as an R \strong{dataframe} or \strong{matrix}, must include the features used in
the \code{smooth_specs} specifications.}

\item{y}{The model target as an R \strong{vector}.}

\item{smooth_specs}{A named list of smooth specifications.}

\item{requires_grad}{A boolean indicating whether or not the \code{x} tensors will need gradients.
This can be turned off for datasets that are only used for evaluation, e.g. validation data.}
}
\description{
Builds \code{wand} compatible \code{torch::tensor_dataset}s
}
