% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_wand_dataset}
\alias{build_wand_dataset}
\title{Convert a matrix to a \code{wand} compatible \code{torch::dataset}.}
\usage{
build_wand_dataset(x, y, smooth_features)
}
\description{
Convert a matrix to a \code{wand} compatible \code{torch::dataset}.
}
