% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_smooths}
\alias{extract_smooths}
\title{Takes a formula, and replaces any smooth specs with the corresponding formula. Returns the
updated formula and the smooth specs extracted from the formula}
\usage{
extract_smooths(formula)
}
\description{
Takes a formula, and replaces any smooth specs with the corresponding formula. Returns the
updated formula and the smooth specs extracted from the formula
}
