% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_var_sequence}
\alias{get_var_sequence}
\title{Generates an evenly spaced sequence across a variable's values}
\usage{
get_var_sequence(var)
}
\arguments{
\item{var}{A variable, numeric or factor.}
}
\value{
An evenly spaced sequence from \code{var} min to max.
}
\description{
Generates an evenly spaced sequence across a variable's values
}
