% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{incant}
\alias{incant}
\title{Retrieve 'magic' attributes from files and directories}
\usage{
incant(path)
}
\arguments{
\item{path}{character vector of files to use magic on}
}
\value{
a \code{tibble} / \code{data.frame} of file magic attributes.
  Specifically, mime type, encoding, possible file extensions and
  type description are returned as colums in the data frame along
  with \code{path}.
}
\description{
Retrieve 'magic' attributes from files and directories
}
\examples{
library(magrittr)
library(dplyr)

system.file("img", package="filemagic") \%>\%
  list.files(full.names=TRUE) \%>\%
  incant() \%>\%
  glimpse()
}

