% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wand-predict.R
\name{predict.wand}
\alias{predict.wand}
\title{Predict from a \code{wand}}
\usage{
\method{predict}{wand}(object, new_data, type = NULL, ...)
}
\arguments{
\item{object}{A \code{wand} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
\item \code{"class"} for hard class predictions.
\item \code{"prob"} for soft class probability predictions.
}}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{wand}
}
\examples{
\donttest{
train <- mtcars[1:20,]
test <- mtcars[21:32, -1]

# Fit
mod <- wand(mpg ~ cyl + log(drat), train)

# Predict, with preprocessing
predict(mod, test)
}

}
