% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{simplemagic_mime_db}
\alias{simplemagic_mime_db}
\title{File extension-to-MIME mapping data frame}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1763 rows and 2 columns.}
\usage{
simplemagic_mime_db
}
\description{
File extension-to-MIME mapping data frame
}
\keyword{datasets}
