% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_smooth_formula}
\alias{validate_smooth_formula}
\title{Validate a formula for a smooth specification.}
\usage{
validate_smooth_formula(formula)
}
\arguments{
\item{formula}{A formula.}
}
\value{
The unchanged formula.
}
\description{
Validate a formula for a smooth specification.
}
