% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wand-package.R
\docType{package}
\name{wand-package}
\alias{wand}
\alias{wand-package}
\title{wand: Fit Wide and Deep Neural Networks}
\description{
TODO
}
\author{
\strong{Maintainer}: Ben Ewing \email{ben.j.ewing@protonmail.com}

}
\keyword{internal}
