% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wand-package.R
\docType{package}
\name{wand}
\alias{wand}
\alias{wand-package}
\title{Retrieve 'Magic' Attributes from Files and Directories}
\description{
The 'libmagic' library provides functions to determine 'MIME' type and other
metadata from files through their "magic" attributes. This is useful when you
do not wish to rely solely on the honesty of a user or the extension on a
file name. It also incorporates other metadata from the mime-db database
\url{https://github.com/jshttp/mime-db}
}
\section{Some important details}{


The header checking is minimal (i.e. nowhere near as comprehensive as \code{libmagic}) but
covers quite a bit of ground. If there are content-check types from
\href{https://github.com/threatstack/libmagic/tree/master/magic/}{magic sources}
that you would like coded into the package, please file an issue and
\emph{include the full line(s)} from that linked \code{magic.tab} that you would like mapped.
}

\author{
Bob Rudis (bob@rud.is)
}
