% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wand-package.r
\docType{package}
\name{wand}
\alias{wand}
\alias{wand-package}
\title{Retrieve 'Magic' Attributes from Files and Directories}
\description{
The 'libmagic' library provides functions to determine 'MIME' type and other
metadata from files through their "magic" attributes. This is useful when you
do not wish to rely solely on the honesty of a user or the extension on a
file name. It also incorporates other metadata from the mime-db database
<https://github.com/jshttp/mime-db>
}
\details{
Based on \code{file} / \code{libmagic} - \url{https://github.com/file/file}
}
\author{
Bob Rudis (@hrbrmstr)
}

