% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_inspection.R
\name{wand_plot_smooths}
\alias{wand_plot_smooths}
\title{Plot smooth functions from a wand model}
\usage{
wand_plot_smooths(wand_fit, seq_length = 250)
}
\arguments{
\item{wand_fit}{A fitted \code{wand} model.}

\item{seq_length}{The maximum number of points used when plotting each smooth. If the smooth is
2d, this represents the number of unique points used for each axis, such that the total number
of points is \code{seq_length*seq_length}.}
}
\value{
A list where each entry is a ggplot of a smooth function. Note that smooths are plotted
after any pre-processing.
}
\description{
Plot smooth functions from a wand model
}
